+++ lib/database/adapter-websql.d.ts
@@ -33,6 +33,7 @@ export declare class _WebSQLStore implements NanoSQLStorageAdapter {
     write(table: string, pk: DBKey | null, data: DBRow, complete: (row: DBRow) => void, skipReadBeforeWrite: boolean): void;
     delete(table: string, pk: DBKey, complete: () => void): void;
     read(table: string, pk: DBKey, callback: (row: DBRow) => void): void;
+    batchRead(table: string, pks: any[], callback: (rows: any[]) => void): void;
     rangeRead(table: string, rowCallback: (row: DBRow, idx: number, nextRow: () => void) => void, complete: () => void, from?: any, to?: any, usePK?: boolean): void;
     drop(table: string, callback: () => void): void;
     getIndex(table: string, getLength: boolean, complete: (index) => void): void;