+++ lib/database/storage.d.ts
@@ -38,7 +38,7 @@ export interface NanoSQLStorageAdapter {
      * @param {() => void} complete
      * @memberof NanoSQLStorageAdapter
      */
-    connect(complete: () => void): void;
+    connect(complete: () => void, error?: (err: Error) => void): void;
     /**
      * Write a single row to the database backend.
      * Primary key will be provided if it's known before the insert, otherwise it will be null and up to the database backend to make one.
@@ -51,7 +51,7 @@ export interface NanoSQLStorageAdapter {
      * @param {boolean} skipReadBeforeWrite
      * @memberof NanoSQLStorageAdapter
      */
-    write(table: string, pk: DBKey | null, data: DBRow, complete: (finalRow: DBRow) => void, skipReadBeforeWrite: boolean): void;
+    write(table: string, pk: DBKey | null, data: DBRow, complete: (finalRow: DBRow) => void, skipReadBeforeWrite: boolean, error?: (err: Error) => void): void;
     /**
      * Read a single row from the database
      *
@@ -60,7 +60,18 @@ export interface NanoSQLStorageAdapter {
      * @param {(row: DBRow ) => void} callback
      * @memberof NanoSQLStorageAdapter
      */
-    read(table: string, pk: DBKey, callback: (row: DBRow) => void): void;
+    read(table: string, pk: DBKey, callback: (row: DBRow) => void, error?: (err: Error) => void): void;
+    /**
+     * Given an arbitrary list of primary keys and a table, get all primary keys.
+     * This method is optional, if it isn't provided then .read() will be called in parallel to perform these kinds of queries.
+     *
+     * @param {string} table
+     * @param {DBKey[]} pks
+     * @param {(rows: DBRow[]) => void} callback
+     * @param {(err: Error) => void} [error]
+     * @memberof NanoSQLStorageAdapter
+     */
+    batchRead?(table: string, pks: DBKey[], callback: (rows: DBRow[]) => void, error?: (err: Error) => void): void;
     /**
      * Read a range of primary keys from a given table.
      * Each row is read asyncrounosuly, so make sure the front end can incriment through the rows quickly.
@@ -75,7 +86,7 @@ export interface NanoSQLStorageAdapter {
      * @param {DBKey} [to]
      * @memberof NanoSQLStorageAdapter
      */
-    rangeRead(table: string, rowCallback: (row: DBRow, idx: number, nextRow: () => void) => void, complete: () => void, from?: any, to?: any, pkRange?: boolean): void;
+    rangeRead(table: string, rowCallback: (row: DBRow, idx: number, nextRow: () => void) => void, complete: () => void, from?: any, to?: any, pkRange?: boolean, error?: (err: Error) => void): void;
     /**
      * Delete a row from the backend given a table and primary key.
      *
@@ -84,7 +95,7 @@ export interface NanoSQLStorageAdapter {
      * @param {() => void} complete
      * @memberof NanoSQLStorageAdapter
      */
-    delete(table: string, pk: DBKey, complete: () => void): void;
+    delete(table: string, pk: DBKey, complete: () => void, error?: (err: Error) => void): void;
     /**
      * Drop an entire table from the backend. (Delete all rows)
      *
@@ -92,7 +103,7 @@ export interface NanoSQLStorageAdapter {
      * @param {() => void} complete
      * @memberof NanoSQLStorageAdapter
      */
-    drop(table: string, complete: () => void): void;
+    drop(table: string, complete: () => void, error?: (err: Error) => void): void;
     /**
      * Get the number of rows in a table or the table index;
      *
@@ -100,14 +111,14 @@ export interface NanoSQLStorageAdapter {
      * @param {(count: number) => void} complete
      * @memberof NanoSQLStorageAdapter
      */
-    getIndex(table: string, getLength: boolean, complete: (index: any[] | number) => void): void;
+    getIndex(table: string, getLength: boolean, complete: (index: any[] | number) => void, error?: (err: Error) => void): void;
     /**
      * Completely delete/destroy the entire database.
      *
      * @param {() => void} complete
      * @memberof NanoSQLStorageAdapter
      */
-    destroy(complete: () => void): any;
+    destroy(complete: () => void, error?: (err: Error) => void): any;
 }
 /**
  * Holds the general abstractions to connect the query module to the storage adapters.
@@ -316,7 +327,8 @@ export declare class _NanoSQLStorage {
     _secondaryIndexRead(table: string, column: string, search: string, callback: (rows: DBRow[]) => void): void;
     /**
      * Get a range of rows from a given table.
-     * The range is in limit/offset form where the from and to values are numbers indicating a range of rows to get.
+     * If usePKs is false the range is in limit/offset form where the from and to values are numbers indicating a range of rows to get.
+     * Otherwise the from and to values should be primary key values to get everything in between.
      *
      * @param {string} table
      * @param {DBKey} from
@@ -324,18 +336,7 @@ export declare class _NanoSQLStorage {
      * @param {(rows: DBRow[]) => void} complete
      * @memberof _NanoSQLStorage
      */
-    _rangeReadIDX(table: string, fromIdx: number, toIdx: number, complete: (rows: DBRow[]) => void): void;
-    /**
-     * Get a range fo rows from a given table.
-     * The range is provided as the rows between two primary key values.
-     *
-     * @param {string} table
-     * @param {*} fromPK
-     * @param {*} toPK
-     * @param {(rows: DBRow[]) => void} complete
-     * @memberof _NanoSQLStorage
-     */
-    _rangeReadPKs(table: string, fromPK: any, toPK: any, complete: (rows: DBRow[]) => void): void;
+    _rangeRead(table: string, from: any, to: any, usePKs: boolean, complete: (rows: DBRow[]) => void): void;
     /**
      * Full table scan if a function is passed in OR read an array of primary keys.
      *